reconfiguração

+++ resetar_docker.sh
@@ -1,6 +1,26 @@
 #!/bin/bash
-docker stop $(docker ps -aq) # Para todos os containers
-docker rm $(docker ps -aq)   #
-docker volume rm $(docker volume ls -q)
-docker image rm $(docker image ls -q)
-docker network rm $(docker network ls -q)
+
+# Confirmar com o usuário
+echo "Este script irá remover todos os containers, imagens, volumes e redes do Docker!"
+read -p "Tem certeza que deseja continuar? (s/n): " resposta
+
+if [ "$resposta" = "s" ] || [ "$resposta" = "S" ]; then
+    echo "Iniciando a limpeza do Docker..."
+
+    docker stop $(docker ps -q)
+
+    docker rm $(docker ps -aq)
+
+    docker rmi $(docker images -q) -f
+
+    docker volume rm $(docker volume ls -q)
+
+    docker network prune -f
+
+    docker system prune -a --volumes -f
+
+    echo "Limpeza do Docker concluída!"
+
+else
+    echo "Operação cancelada."
+fi
